+++ extension/src/time-sheets/storage.ts
@@ -1,15 +1,11 @@
 import { addDays, DateOnly, toDateOnlyKey } from '../types/dates'
-import { isTimeSheet, TimeSheet } from '../types/time-sheet'
+import { TimeSheet } from '../types/time-sheet'
 
 export const getTimeSheet = async (weekMondayDate: DateOnly): Promise<TimeSheet> => {
   const key = `timesheet-${toDateOnlyKey(weekMondayDate)}`
   const result = await chrome.storage.local.get([key])
 
-  if (result[key] && isTimeSheet(result[key])) {
-    return result[key]
-  }
-
-  return {
+  const fallbackTimeSheet = TimeSheet.parse({
     dates: {
       monday: weekMondayDate,
       tuesday: addDays(weekMondayDate, 1),
@@ -20,5 +16,17 @@ export const getTimeSheet = async (weekMondayDate: DateOnly): Promise<TimeSheet>
       sunday: addDays(weekMondayDate, 6),
     },
     timeCards: [],
+  })
+
+  if (!result[key]) {
+    return fallbackTimeSheet
   }
+
+  const parseResult = TimeSheet.safeParse(result[key])
+
+  if (parseResult.success) {
+    return parseResult.data
+  }
+
+  return fallbackTimeSheet
 }